Added functionality to generate training files from GIB
